Refactor schema organization by categorizing into 'api' and 'tools' packages.
