+++ src/memlord/schemas/tools/delete.py
