+++ src/memlord/schemas/tools/store.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class StoreResult(BaseModel):
+    name: str
+    created: bool
